Fix for the serialtaskqueuechain_t crashes.
It appears to be true that cppunit is not thread-safe.  Tests
using it crash about once every 5 to 20 times run with heap
corruption.  Sometimes it is possible to see that this is coming
from unprotected uses of container insertion/deletion/iteration
inside of cppunit code.

To compensate the serialtaskqueuechain_t test has been rewritten
to not use cppunit.  Amusingly this makes the test a great deal
faster, cppunit has quite the overhead.